Fe/feature/market detail

+++ backend/src/main/java/com/zzol/sizzang/review/dto/request/ReviewAddReq.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class ReviewAddReq {
 
     private int userCode;
-    private int stCode;
+    private Long stCode;
     private int puCode;
     private String reTitle;
     private String reContent;
